Revert "Option to opt-out of having auto-scale-group controller elastic IP." This reverts commit 6273dddf35fea79a2af9eaba9fcad492ebbd0aab.
Summary:
EC2 instances need a public IP in order to have outbound internet access
(ie in order for the replies to come back) unless setup with a proper private subnet
with routes through to a NAT gateway.

This commit didn't consider this and ended up with controller unable to reach the internet
eg for apt-get updates etc.

Another attempt could be made for a function eg
createPrivateInstance (would create an EC2 instance, not create an EIP, would require a private subnet,
should require the controller_internal_dns_zone logic as present in the reverted commit
and should be tested to have outbound internet access)

Reviewers: timd, bacek

Subscribers: #helix_omniscient

Differential Revision: http://phab.helixta.com.au/D35380